@@ -49,7 +49,7 @@
     <properties>
         <projectOwner>Red Hat, Inc.</projectOwner>
         <projectEmail>nos-devel@redhat.com</projectEmail>
-        <javaVersion>1.8</javaVersion>
+        <javaVersion>11</javaVersion>
         <test-forkCount>1</test-forkCount>
         <datastaxVersion>3.7.2</datastaxVersion>
         <cassandraUnitVersion>3.11.2.0</cassandraUnitVersion>
