@@ -15,8 +15,8 @@
  */
 package org.commonjava.storage.pathmapped.pathdb.jpa.model;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 import java.io.Serializable;
 import java.util.Objects;
 
