@@ -17,10 +17,10 @@ package org.commonjava.storage.pathmapped.pathdb.jpa.model;
 
 import org.commonjava.storage.pathmapped.model.Reclaim;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.util.Date;
 import java.util.Objects;
 
